py-scripts/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "pandas>=2.2.3",
    "pymongo>=4.12.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
